requirements.txt
alembic==1.4.3
atomicwrites==1.4.0
attrs==20.3.0
cffi==1.14.4
click==7.1.2
colorama==0.4.4
cryptography==3.3.1
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.10.0
packaging==20.8
pluggy==0.13.1
py==1.10.0
pycparser==2.20
PyJWT==2.0.0
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
webargs==7.0.1
Werkzeug==1.0.1
